src/freelance_jobs/pyproject.toml
[project]
name = "freelance-jobs"
version = "0.1.0"
description = "Freelance Job Search Hook for Discord Bot"
dependencies = [
    "bot-common",
    "discord-webhook>=1.3.0",
    "python-dotenv>=1.0.0",
]

[tool.uv.sources]
bot-common = { workspace = true }
